+++ R/vrt-set-nodata.R
@@ -33,14 +33,7 @@ vrt_set_nodata.default <- function(x, ...) {
   )
 }
 
-#' @export
-#' @rdname vrt_set_nodata
-vrt_set_nodata.vrt_block <- function(
-  x,
-  nodatavalue,
-  nodata = nodatavalue,
-  band_idx = NULL
-) {
+apply_set_nodata_xml <- function(x, nodatavalue, nodata, band_idx) {
   v_assert_type(nodatavalue, "nodatavalue", "numeric", nullok = FALSE)
   v_assert_type(nodata, "nodata", "numeric", nullok = FALSE)
   v_assert_type(
@@ -58,19 +51,12 @@ vrt_set_nodata.vrt_block <- function(
     ),
     function(v, n) {
       if (length(v) > 1) {
-        v_assert_length(
-          v,
-          n,
-          length(x$assets),
-          nullok = FALSE
-        )
+        v_assert_length(v, n, length(x$assets), nullok = FALSE)
       }
     }
   )
 
-  # Get the VRT XML
   vrt_xml <- xml2::read_xml(x$vrt)
-
   bands <- xml2::xml_find_all(vrt_xml, ".//VRTRasterBand")
   if (is.null(band_idx)) {
     band_idx <- seq_along(bands)
@@ -79,27 +65,50 @@ vrt_set_nodata.vrt_block <- function(
   purrr::pwalk(
     list(band = bands[band_idx], nodatavalue = nodatavalue, nodata = nodata),
     function(band, nodatavalue, nodata) {
-      # Remove existing NoDataValue elements
       reset_element(band, "NoDataValue", nodatavalue)
-
       band_src <- xml2::xml_find_first(band, ".//ComplexSource")
-      # Remove existing NODATA elements
       reset_element(band_src, "NODATA", nodata)
     }
   )
-  out_vrt <- fs::file_temp(
-    tmp_dir = getOption("vrt.cache"),
-    ext = "vrt"
-  )
-  # Save the modified VRT XML back to the file
-  xml2::write_xml(vrt_xml, out_vrt)
 
+  out_vrt <- fs::file_temp(tmp_dir = getOption("vrt.cache"), ext = "vrt")
+  xml2::write_xml(vrt_xml, out_vrt)
+  out_vrt
+}
+
+#' @export
+#' @rdname vrt_set_nodata
+vrt_set_nodata.vrt_block <- function(
+  x,
+  nodatavalue,
+  nodata = nodatavalue,
+  band_idx = NULL
+) {
+  out_vrt <- apply_set_nodata_xml(x, nodatavalue, nodata, band_idx)
   build_vrt_block(
     out_vrt,
     pixfun = x$pixfun,
     maskfun = x$maskfun,
     warped = x$warped,
     is_remote = x$is_remote
+  )
+}
+
+#' @export
+#' @rdname vrt_set_nodata
+vrt_set_nodata.vrt_stack <- function(
+  x,
+  nodatavalue,
+  nodata = nodatavalue,
+  band_idx = NULL
+) {
+  out_vrt <- apply_set_nodata_xml(x, nodatavalue, nodata, band_idx)
+  build_vrt_stack(
+    out_vrt,
+    n_items = x$n_items,
+    maskfun = x$maskfun,
+    pixfun = x$pixfun,
+    warped = inherits(x, "vrt_stack_warped")
   )
 }